Enable multisim for modular simulator
This enables the general framework for multisims in modular simulator.

It also extracts the calculation of the signal communication period out
of do_md into a helper function to avoid duplication.

Add tests to show equivalence of multi-sim with MD and VV with and
without modular simulator.

Multi-sim tests now use GTEST_SKIP so there is feedback when a test
cannot be run.

TODO test replica exchange with the modular simulator

Refs #3433