gmxapi reading and writing of coordinates from tpr files.
Add a light-weight buffer description struct for coordinates to support
external data access such as Eigen and Numpy. Implement the
Python Buffer Protocol.

Allow no-copy reads from `t_state` through gmxapi::StructureSource.

Add TprBuilder to ::gmxapicompat. Setting coordinates for output writing
requires a copy because a `t_state` is required for TPR output and
`t_state` has internal storage.

There isn't a good way to get the full memory layout for the coordinates
storage in `t_state`, but fortunately the coordinates appear to be
stored contiguously. Many checks are added to confirm this and to catch
broken assumptions if the storage changes in the future.

Refs #3295