Add nblib gpu non-bonded force calculator
Add the interface for non-bonded force calculation in nblib. One
compute function takes host buffers and does transfer and reordering
each step. Another takes device buffers so that a gpu resident step
can be implemented. Full support for a gpu resident step requires
gpu integration, which is not implemented as part of this MR.