Use Update module to freeze fully frozen atoms
In theory, fully frozen atoms should be dealt with where their
coordinates are touched. In practice, this is easy to overlook and
e.g. pressure coupling can move atos relying on update to freeze
them after. This moves the responsibility of freezing atoms to Update
even if they are fully frozen. The resulting behaviour is more robust.

Refs #3075