Remove gmxapi sample_restraint tests to fix CI failures
These tests depend on cloning one or more external
repositories at runtime, which presently fails because
there have been changes in the repositories outside
Gromacs. While that might be possible to fix, it also
means the tests are non-reproducible and rather depend
on the particular state of other git repositories when
run.

For now this is causing all CI jobs to fail, so I have
disabled the tests in question until somebody has time
to look into a proper long-term solution.