Move and update GROMACS contributor and leader lists
The list of contributors has been moved to the conventional location
(a top-level AUTHORS file) and no longer written to the terminal
output for every run. There's no need to tell every user of every tool
who all the contributors were and who owns the copyrights or what the
license is. In particular, the copyright statement pertains only to
the source and need only appear there. Since the license is free, usage
is not restricted and there is no need to inform users of that, just
like almost every other command-line tool does not so inform users.

Examples of the kinds of contributions that we wish to recognize have
been made. Active contributors are now distinguished from past ones.

Some past contributors' names have been added based on "git shortlog -e
--summary --numbered".

The leader list now reflects current activity.