Make automatic release builds
Adds a script to run the complete automated release workflow as the last
step of the release workflow. This step is only ever run manually, and
only uploads to the correct location when running a full release build
(with GROMACS_RELEASE == true) on one of the branches we do releases
from (master, release-2022).