Don't ignore pinning failures when this was requested
If pinning fails on any thread across the simulation, a harmless-looking
note will be issued in the log and mdrun will always continue.
When mdrun -pin on requests thread pinning, failures to set affinities
should not be ignored as this can lead to major performance loss.

This change improves error reporting to inform users when setting
affinities failed and honors the manual pinning request by terminating
the run if affinity setting failed.

Fixes #4350