Don't prevent esoteric function types in free energy if they are unperturbed
Previously, our topology generation code would check all
listed interactions, and issue a fatal error whenever 
a function type that cannot be perturbed was detected while
having free energy calculation enabled, even if that
function type was not perturbed.

This changes the logic to first check if the function
type is actually perturbed (different state A/B parameters),
and if the type itself is not perturbed it will be allowed
even if free energy is used for other interactions.

Fixes #3691.