Infrastructure to modify charges during a simulation with a lambda parameter
This is a sketch on infrastructure that allows to change charges for specific sets of atoms.

Design is based on a managing object that does the modifications and hands out handlers to objects that can request such a modification - a similar design pattern to LocalAtomSet and LocalAtomSetManager.

The charge modifier is accessible to MdModules via an additional callback.

Basic infrastructure for #4273 , relates to #4014 #3956