Fix SFINAE on SIMD load for smoother type deduction
If the SFINAE goes on the return type then template deduction can work
and no ambiguity can ever arise. The function parameter for load() is
always of arithmetic type (by construction of SimdTraits) and the
return value is either SimdReal or real, which also determines T if
not given explicitly.

This permits simplifying all the cases that formerly explicitly named
the function template in order to avoid needing to use template
deduction that didn't work. Now the return type does the deduction and
is always correct. Both before and after this change, the user can
still make the error of intending to load a SIMD real, in fact
assigning to a non-SIMD real, and so actually loading a non-SIMD real.

Removed some unnecessary use of std::remove_const and a function that
was both unnecessary and unused.