Add option to read additional itp files into pdb2gmx
This allows users to specify ligand itp files during topology generation
that are then automatically included in the final topology (if they are
valid). The feature works the same way as for the inclusion of solvent
itp files, and is tested that way.

Relies on ability to read in several input files on command line.