Use TimeControl properly
Add helper function to clear previously set start and end times when the
tool exits, to ensure they are correctly cleaned up. This caused the bug
before when a tool would have set a time value, but then runs again in
the same executable without the same value being set.