Link utility privately in tests
Almost all or perhaps all tests currently have at least a transitive
dependence on utility. This change makes that explicit. It is likely
that as more libraries become object targets, some of these
dependencies will be shown to be spurious, at which point they can
be replaced by linking the correct dependency.

Related to #3288