Make a cuda_version OBJECT target in utility
This demonstrates one part of the path forward for updating cuda
library handling. In order to have each module be an object target,
each module must also take care of its GPU dependencies, instead of
leaving that to libgromacs. This demonstrates the trivial case
where the cu file has no dependence on the object it is linked in
to.

Relates to #3288