'Bump minimum cmake version to 3.17 to support cuda restructuring
Cmake 3.17 was release on 20 March 2020. The default version of cmake
in Ubuntu 22.04 is 3.22.1, so there should not be a problem there.

This is necessary because cmake 3.17 introduces find_package(CUDAToolkit),
which is a major step forward in usability over find_package(CUDA) or
enable_language(CUDA). In particular, with the library structure before
this change is introduced, it is not possible to do a CUDA build
on a system with no GPU but an installed CUDA toolkit, which was
possibled before the introduction of enable_language(CUDA), in
!2994. This will also make packaging cuda code much easier, as
without this change, it would be necessary to handle the different
variables that are set by the old find_package(CUDA) and the newer
find_package(CUDAToolkit). This work would also be somewhat pointless
busywork since surely we would anyway bump to at least cmake 3.17 after
the release.

Relates to #3288