[wip] Clarify module dependencies in `src/gromacs`.
Remove the catch-all `legacy_modules` target and stop including
`src/` in all of the include paths. Require `target_link_libraries`
to get the interface headers for specific modules.
* Move module headers such that they are uniquely exposed to
  `gromacs/modulename` by a single target.
* Update the `target_link_libraries` for dependent modules (though many
  will be overlooked until the sources are compartmentalized for modules
  currently defined as INTERFACE type).

There is not a good way to restrict access to the private module headers
while all of the sources are aggregated monolithically into the
libgromacs target. The best we can do is to use relative paths in the
`#include` statements when nested sources need to access private module
headers. Fortunately, it will be relatively easy to locate and clean up
dependencies of the form `#include "../<...>"`.

Refs #3288.