Add optimized fast_float3 implementation for CDNA2
HIP's float3 is defined as a struct of 3 fields, the compiler is not aware of its vector nature
hence it is not able to generate optimal packed math instructions (v_pk_...). This new type is
defined as struct of float2 (x, y) and float (z) so packed math can be used for x and y.

Refs: #4593