Reduce scope of some gpu_utils headers
Forward declare DeviceContext and DeviceStream instead of including
in several headers. Remove include of gputraits.h from several headers.
In a few cases this is achieved by using templating in exposed
headers and instantiating in the source. This change will make
it easier to have interfaces where the dataflow is more clear,
especially in the sycl case where DeviceBuffer is a class and
not a using declaration.