Avoid violating ODR with internal structures
`gmx::AtomPair` is defined in both `src/gromacs/mdlib/lincs.cpp` and
`src/gromacs/mdlib/lincs_gpu.h`. Moving it into an anonymous namespace
in the former file since it's used internally.

The structures are pretty much identical (just a pair of int's), so the
violation is likely harmless. Still, better get rid of it.

Compilation error reported in
https://gromacs.bioexcel.eu/t/error-installing-gromacs-2022-c-one-definition-rule/5322.