Decouple applied-forces tests from ForceProviderInput contents and constructor
Up to now there was just a single constructor for ForceProviderInput that
would require all parameters that the class holds. This constructor is fine
for using it in the MD loop, where all those parameters actually need to be passed.
However, uing that full constructor in test code for a force provider
(as e.g. in applied-forces/tests/electricfield.cpp) requires to change
that test each time the constructor is extended in the future, although
such changes will likely be unrelated to the test. Adding a basic
constructor with just the minimal required arguments (commrec) solves
that issue. Parameters that the individual tested force provider needs
can simply be passed by setting respective (public) ForceProviderInput
variables.