CI: Add scripts and configs for AppleSilicon testing
GitLab runner does not support docker on macOS, only plain shell, so we
use a bash script to install dependencies.

GitLab caches it locally, so bootstrapping is not repeated
(even from scratch, it takes around 5 minuts).
The cache uses the checksum of the bootstrapping
script as the key, so multiple versions would be able to co-exist.

We are using system-wide Apple Clang, no way to pin its version.