Add external {fmt} library while waiting for c++20
{fmt} is an outstanding text formatting library that
is both C++-centric (meaning strings), it gets rid
of the different format specifiers for different
integer widths, it is much faster - and it will catch
many formatting errors already at compile-time.
This library is part of the C++20 standard, but
until we require C++20 support we can use it as an
external library and only change the namespace from
fmt:: to std:: next year.