Reduce algorithmic complexity of calcVerletBufferDrift()
Use unordered_map for obtained the list of Verlet buffer atom types.
This reduces the complexity from O(#atoms * #types) to O(#atoms).
Reduced the resolution of mass, charge and constraint length of
the Verlet buffer atom types from real to int16_t. The resolution
is set as 0.01*RMS. This should significantly reduce the number
of types in systems where the charges for each atom in each different
molecule are different, thereby reducing the O(#types^2) operations
cost.

Fixes #4892